+++ saphyr/src/lib.rs
@@ -55,7 +55,7 @@ pub use crate::annotated::{
 };
 pub use crate::emitter::{EmitError, YamlEmitter};
 pub use crate::loader::{LoadError, LoadableYamlNode, YamlLoader};
-pub use crate::scalar::{Scalar, ScalarOwned};
+pub use crate::scalar::{parse_core_schema_fp, Scalar, ScalarOwned};
 pub use crate::yaml::{Mapping, Sequence, Yaml, YamlIter};
 pub use crate::yaml_owned::{MappingOwned, SequenceOwned, YamlOwned, YamlOwnedIter};